\name{gen_ints}
\docType{data}
\alias{gen_ints}
\alias{i}
\alias{j}
\alias{k}
\title{Genome Intervals examples}
\description{
    Toy examples for testing functions and running examples of the package \code{genomeIntervals}.
}
\usage{ data(gen_ints) }
\format{Two \code{Genome_intervals_stranded} objects, \code{i} and \code{j}, without inter-base intervals and a third one, \code{k}, with.}
\keyword{datasets}
