% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Genome_intervals-package.R
\name{Genome_intervals deprecated functions}
\alias{Genome_intervals deprecated functions}
\alias{seq_name}
\alias{seq_name,Genome_intervals-method}
\alias{seq_name<-}
\alias{seq_name<-,Genome_intervals-method}
\title{The following function have been deprecated:
\itemize{
\item \code{seq_name}
\item \code{seq_name<-}
}}
\description{
\itemize{
\item The \code{seq_name} and \code{seq_name<-} accessor functions have been
replaced by the more generic \code{seqnames} and \code{seqnames<-} accessor
functions, respectively.
}
}

