% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Genome_intervals-coercion-methods.R
\name{genomeIntervals coercion methods}
\alias{as}
\alias{as,Genome_intervals,GRangesList-method}
\alias{as,Genome_intervals,RangedData-method}
\alias{as,Genome_intervals-method}
\alias{coerce,Genome_intervals,GRangesList-method}
\alias{coerce,Genome_intervals,RangedData-method}
\alias{genomeIntervals coercion methods}
\title{Coercion methods of the genomeIntervals package}
\usage{
\S4method{as}{Genome_intervals}(from,to)
}
\arguments{
\item{from}{An object of class \code{\linkS4class{Genome_intervals}}}

\item{to}{a character string; either RangedData or GRangesList}

\item{x}{An object of class \code{\linkS4class{Genome_intervals}}}
}
\value{
\describe{
\item{coerce}{ A \code{\linkS4class{RangedData}} or
\code{\linkS4class{GRangesList}}
containing the result of the coercion.  }}
}
\description{
\describe{
\item{coerce}{ This method allows to coerce a
\code{\link[genomeIntervals:Genome_intervals_stranded-class]{genomeIntervals
object}} into a \code{\link[IRanges:RangedData-class]{RangedData object}} or
\code{\linkS4class{GRangesList}} object.}}
}
\examples{
\dontrun{
annot<-readGff3(system.file("extdata",
                            "Dmel-mRNA-exon-r5.52.gff3",
                            package="RnaSeqTutorial")
gAnnot<-as(annot,"RangedData") type(annot)
}
}
\author{
Nicolas Delhomme
}
\seealso{
\itemize{
\item \code{\link[genomeIntervals:Genome_intervals_stranded-class]{genomeIntervals
object}}
\item \code{\link[genomeIntervals:readGff3]{readGff3 function}}
}
}

